+++ mobile/shared/modules/geckoview/GeckoViewContent.sys.mjs
@@ -38,6 +38,7 @@ export class GeckoViewContent extends GeckoViewModule {
       "GeckoView:UpdateInitData",
       "GeckoView:ZoomToInput",
       "GeckoView:IsPdfJs",
+      "GeckoView:GetWebCompatInfo",
     ]);
   }
 
@@ -272,6 +273,9 @@ export class GeckoViewContent extends GeckoViewModule {
       case "GeckoView:ContainsFormData":
         this._containsFormData(aCallback);
         break;
+      case "GeckoView:GetWebCompatInfo":
+        this._getWebCompatInfo(aCallback);
+        break;
       case "GeckoView:RequestAnalysis":
         this._requestAnalysis(aData, aCallback);
         break;
@@ -415,6 +419,36 @@ export class GeckoViewContent extends GeckoViewModule {
         }
         break;
       }
+    }
+  }
+
+  async _getWebCompatInfo(aCallback) {
+    if (
+      Cu.isInAutomation &&
+      Services.prefs.getBoolPref(
+        "browser.webcompat.geckoview.enableAllTestMocks",
+        false
+      )
+    ) {
+      const mockResult = {
+        devicePixelRatio: 2.5,
+        antitracking: { hasTrackingContentBlocked: false },
+      };
+      aCallback.onSuccess(JSON.stringify(mockResult));
+      return;
+    }
+    try {
+      const actor =
+        this.browser.browsingContext.currentWindowGlobal.getActor(
+          "ReportBrokenSite"
+        );
+      const info = await actor.sendQuery("GetWebCompatInfo");
+
+      // Stringify to convert potential non-ASCII
+      // characters in the returned web compat info map.
+      aCallback.onSuccess(JSON.stringify(info));
+    } catch (error) {
+      aCallback.onError(`Cannot get web compat info, error: ${error}`);
     }
   }